st/mesa: use proper color-renderable check
Any color-renderable texture can be bound to an FBO, so we need to
specify PIPE_BIND_RENDER_TARGET when picking the texture-format in
that case.

Unfortunately, the old check for this was rather lacking (and even
incorrect; internalformat cannot be the values 3 or 4, only the
format can be). This lead us to depend on the driver to support
rendering any format being picked that fell through the cracks.

So let's use this new and shiny function to figure out if the
format really are color-renderable instead.