mesa/fbo: only consider api-format when validating fbo
texImg->TexFormat is chosen by the driver based on the user-input,
but is not the actual api-format. However, color-renderable is
defined as a function of the api-format itself. So it's stricly
more correct to consider only the api-format, and not the
driver-chosen format when deciding if an attachment is
color-renderable or not.

For OpenGL, this means the following changes:
- GL_RGB9_E5 is no longer color-renderable. It shouldn't be in the
  first place, according to both OpenGL 4.5 and
  EXT_texture_shared_exponent.

Also OpenGL ES gets rid of some incorrectly supported formats:
- GL_RGBA32F and GL_RGBA16F. There's no extension that makes
  32-bit floating-point textures color-renderable in OpenGL ES 2.0,
  and we don't support EXT_color_buffer_half_float, which
  introduces color-renderable 16-bit floating point formats.
- A ton of other formats, like integer textures and whatnot are
  no longer color-renderable. The OpenGL ES specifications are
  pretty strict when it comes to what formats are color-renderable.