main: only consider format-type
There's no wording in the OpenGL specification that indicates that
querying GL_RED_BITS, GL_GREEN_BITS or GL_BLUE_BITS should only
return the bit-sizes for complete framebuffers.

So instead of checking for a valid color format, let's just check
if the format is a color format or not.